Add files via upload
The described program takes an input of .dta (STATA 69) files that are generated by the Gamry Analyst software provided by Gamry instruments. With prompts for the required constants (concentration, E0', and internal resistance value) and the desired electron rate transfer to be calculated, the program finds the heterogeneous rate transfer constants for a standard 0.0707 cm2 (surface area) working electrode and the diffusion constant. You receive a plot and the best fit line details of all requested data. All relevant peak data and final values are subsequently output to a .CSV for your convenience.