Add files via upload
This script processes .DTA (STATA 69) files that are exported by Gamry Analyst, a product of Gamry instruments, to produce the diffusion coefficient and heterogeneous electron transfer constant rates. Once files are saved in the same directory, the script will import all .DTA files it can and provide the ability to exclude specific files. You are prompted for the required constants and the values are calculated assuming a standard 0.0707 cm2 electrode surface area. All results (CV peaks, Do, Ko) are displayed as the requisite graph and all data and constants are exported to a .CSV file for your convenience.