Add files via upload
These files were made to use for the Gamry analyst .DTA (STATA 69) file format output. 

"CV_Kinetics_Analysis" will generate graphs of the Do and selected Ko of the provided .DTA files. The graphs display the pertinent data points, line of best fit (and equation), and fitting R^2 value. The graphs are saved as .PNG. All data for CV peaks and final data are saved in a data frame which is exported as a .CSV for future reference. 

"CV_workup" will generate a plot of a CV with the scan rate, concentration, and sweep direction displayed. The plot is saved as a .PNG. The peak data and E1/2 is saved in a .CSV file for the user.