Add files via upload
This program was made for cropping large amounts of images. The program prompts the user for a file extension type and pulls all of the associated files into a list, with the option to remove any. All files are then cropped to a user-specified pixel dimension centered around middle of the image. The file is then saved as a "cropped" version with the same file extension.