Add files via upload
Program to process data from Cary 60 Spectrometer Kinetics scan program. One or more .CSV files are accepted with user input for the desired wavelength(s), molar absorptivity(s), and cuvette length. The program then finds the kinetic rate law (0th, 1st, or 2nd) that applies to each prompted wavelengths in the file(s). The output is a csv with containing a table of data for each file to recreate the graphs if desired. A final CSV is also output with columns of Order | Wavelength | File