Add files via upload
This program takes an excel file output from Arbin potentiostats and outputs an excel file of pertinent data and statistics such as coulombic, voltaic, and energy efficiencies as well as area specific resistance and state of charge. Plots are also made based on user prompts of the desired data. The program is designed to be used to work up potentiostatic cycling paired with linear sweep amperometery to determine discharge power.