Add files via upload
This program was written to process cyclic voltammetry experiments conducted on Biologic instruments. The .mpr output files are directly processed. The program plots the CVs either as individual files, all of the selected files, or both with the scan rate as the title or legend entry, respectively. The peak values are highlighted and tabulated as well as the initial scan direction (reducing or oxidizing). The data (all scan points and peaks) is exported to a single excel sheet with an additional column for current density. 
The voltage of the scan is adjusted for internal resistance. 
The reference voltage is established by including a scan of a ferrocene reference that has either "ferrocene" of "fc" in the file name. The file is automatically processed to fine the E1/2 of the ferrocene and set that as the 0 V value for the other CV files.