Update LucaCVs.py
The biologic software that outputs the data generates a .mpr file. This is converted to .csv to work up the data. All data (potential, current, coulombs passed, peak locations, etc.) is stored in a .xlsx file. The user can further choose to plot the provided data by individual scan rate values or as an overlay of the second scans of each scan rate provided. All potentials are referenced vs Fc/Fc+ as determined by a provided reference scan file containing Fc or ferrocene in the file name.